add new mocks

+++ Landmarky/ViewModel/DiaryViewModel.swift
@@ -8,8 +8,14 @@ class DiaryViewModel {
 
     @MainActor
     func fetchTrips(modelContext: ModelContext) {
-        let descriptor = FetchDescriptor<Trip>(sortBy: [SortDescriptor(\.startDate, order: .reverse)])
         error = nil
+
+        if Constants.showsMockData {
+            trips = Mock.MockTrips.data.sorted { $0.startDate > $1.startDate }
+            return
+        }
+
+        let descriptor = FetchDescriptor<Trip>(sortBy: [SortDescriptor(\.startDate, order: .reverse)])
         do {
             trips = try modelContext.fetch(descriptor)
         } catch {
@@ -19,6 +25,11 @@ class DiaryViewModel {
 
     @MainActor
     func deleteTrip(_ trip: Trip, modelContext: ModelContext) {
+        if Constants.showsMockData {
+            trips.removeAll { $0.id == trip.id }
+            return
+        }
+
         modelContext.delete(trip)
         do {
             try modelContext.save()
